[refactor] 레스토랑 리스트 섹션 컴포넌트 분리
