+++ src/components/RestaurantListSection/RestaurantList/ListItem/RestaurantInfo/index.tsx
